[박서연] sprint4

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -1,11 +1,10 @@
 package com.sprint.mission.discodeit.service.basic;
 
-import com.sprint.mission.discodeit.dto.BinaryContentRequestDto;
+import com.sprint.mission.discodeit.dto.binaryContent.BinaryContentRequest;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.service.BinaryContentService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,14 +13,18 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
-@Scope("singleton")
 @RequiredArgsConstructor
 public class BasicBinaryContentService implements BinaryContentService {
     private final BinaryContentRepository binaryContentRepository;
 
     @Override
-    public BinaryContent create(BinaryContentRequestDto binaryContentRequestDto) {
-        return binaryContentRepository.save(new BinaryContent(binaryContentRequestDto.data()));
+    public BinaryContent create(BinaryContentRequest binaryContentRequest) {
+        return binaryContentRepository.save(new BinaryContent(
+                binaryContentRequest.fileName(),
+                (long) binaryContentRequest.data().length,
+                binaryContentRequest.contentType(),
+                binaryContentRequest.data()
+                ));
     }
 
     @Override
@@ -37,7 +40,7 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     public List<BinaryContent> findAllByIdIn(List<UUID> binaryContentIds) {
-        return Optional.ofNullable(findAll().stream()
+        return Optional.of(findAll().stream()
                 .filter(binaryContent -> binaryContent.containsId(binaryContentIds))
                 .toList())
                 .orElseThrow(() -> new NoSuchElementException("[ERROR] 존재하지 않는 상태입니다."));
@@ -45,7 +48,12 @@ public class BasicBinaryContentService implements BinaryContentService {
 
     @Override
     public void delete(UUID binaryContentId) {
-        binaryContentRepository.existsById(binaryContentId);
+        if (binaryContentId == null) {
+            return;
+        }
+        if (!binaryContentRepository.existsById(binaryContentId)) {
+            throw new NoSuchElementException("[ERROR] 존재하지 않는 상태입니다.");
+        }
         binaryContentRepository.delete(binaryContentId);
     }
 }
